max value commit

+++ src/main/java/com/it_academy/practice/junit_basics/Calkulator/Calculator.java
@@ -44,7 +44,7 @@ public class Calculator {
             }
             case '/':
 
-                    return (float) a / b;
+                    return  a / b;
 
             case '*': {
                 return a * b;
